Refactor code, move parser tests out into a folder
