+++ crates/percival/src/errors.rs
@@ -1,0 +1,81 @@
+//! Module for human-readable error handling with Ariadne.
+
+use ariadne::{Color, Fmt, Label, Report, ReportKind, Source};
+use chumsky::prelude::*;
+
+/// Format parser errors into a human-readable message.
+pub fn format_errors(src: &str, errors: Vec<Simple<char>>) -> String {
+    let mut reports = vec![];
+
+    for e in errors {
+        let e = e.map(|tok| tok.to_string());
+        let report = Report::build(ReportKind::Error, (), e.span().start);
+
+        let report = match e.reason() {
+            chumsky::error::SimpleReason::Unclosed { span, delimiter } => report
+                .with_message(format!(
+                    "Unclosed delimiter {}",
+                    delimiter.fg(Color::Yellow)
+                ))
+                .with_label(
+                    Label::new(span.clone())
+                        .with_message(format!(
+                            "Unclosed delimiter {}",
+                            delimiter.fg(Color::Yellow)
+                        ))
+                        .with_color(Color::Yellow),
+                )
+                .with_label(
+                    Label::new(e.span())
+                        .with_message(format!(
+                            "Must be closed before this {}",
+                            e.found()
+                                .unwrap_or(&"end of file".to_string())
+                                .fg(Color::Red)
+                        ))
+                        .with_color(Color::Red),
+                ),
+            chumsky::error::SimpleReason::Unexpected => report
+                .with_message(format!(
+                    "{}, expected {}",
+                    if e.found().is_some() {
+                        "Unexpected token in input"
+                    } else {
+                        "Unexpected end of input"
+                    },
+                    if e.expected().len() == 0 {
+                        "end of input".to_string()
+                    } else {
+                        e.expected()
+                            .map(|expected| match expected {
+                                Some(expected) => expected.to_string(),
+                                None => "end of input".to_string(),
+                            })
+                            .collect::<Vec<_>>()
+                            .join(", ")
+                    }
+                ))
+                .with_label(
+                    Label::new(e.span())
+                        .with_message(format!(
+                            "Unexpected token {}",
+                            e.found()
+                                .unwrap_or(&"end of file".to_string())
+                                .fg(Color::Red)
+                        ))
+                        .with_color(Color::Red),
+                ),
+            chumsky::error::SimpleReason::Custom(msg) => report.with_message(msg).with_label(
+                Label::new(e.span())
+                    .with_message(format!("{}", msg.fg(Color::Red)))
+                    .with_color(Color::Red),
+            ),
+        };
+
+        let mut buf = vec![];
+        report.finish().write(Source::from(&src), &mut buf).unwrap();
+        reports.push(std::str::from_utf8(&buf[..]).unwrap().to_string());
+    }
+
+    reports.join("\n")
+}
