+++ crates/percival/src/lib.rs
@@ -5,4 +5,5 @@
 
 pub mod ast;
 pub mod codegen;
+pub mod errors;
 pub mod parser;
